pyproject.toml
[project]
name = "lc-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "langchain-openai>=0.1",
  "python-dotenv>=1.0",
  "requests>=2.31",
  "beautifulsoup4>=4.12",
  "tavily-python>=0.7",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
